@@ -1,39 +1,41 @@
 {
-  "name": "fuma-animate",
+  "name": "fuma-lofi",
   "version": "0.1.0",
   "private": true,
   "scripts": {
-    "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "lint": "next lint",
-    "generate:music": "node ./music/generate.js"
+    "dev": "vite",
+    "build": "vite build",
+    "preview": "vite preview",
+    "generate:music": "node music/generate.mjs"
   },
   "dependencies": {
+    "@fontsource-variable/inter": "^5.0.17",
+    "@fontsource/inter": "^5.0.17",
     "@radix-ui/react-dialog": "^1.0.5",
     "@radix-ui/react-popover": "^1.0.7",
     "@react-spring/three": "^9.7.3",
     "@react-three/drei": "^9.56.5",
     "@react-three/fiber": "^8.12.0",
+    "@vitejs/plugin-react": "^4.2.1",
     "cva": "npm:class-variance-authority@^0.7.0",
     "framer-motion": "^10.9.2",
     "music-metadata": "^7.14.0",
-    "next": "14.1.3",
-    "react": "^18",
-    "react-dom": "^18",
+    "react": "^18.2.0",
+    "react-dom": "^18.2.0",
     "shadergradient": "^1.2.5",
     "tailwind-merge": "^2.2.1",
     "three": "^0.150.1",
-    "three-stdlib": "^2.21.6"
+    "three-stdlib": "^2.21.6",
+    "vite": "^5.1.6"
   },
   "devDependencies": {
-    "@types/node": "^20",
-    "@types/react": "^18",
-    "@types/react-dom": "^18",
+    "@types/node": "^20.11.25",
+    "@types/react": "^18.2.7",
+    "@types/react-dom": "^18.2.7",
     "@types/three": "^0.150.0",
     "autoprefixer": "^10.0.1",
-    "postcss": "^8",
-    "tailwindcss": "^3.3.0",
-    "typescript": "^5"
+    "postcss": "^8.4.35",
+    "tailwindcss": "^3.3.7",
+    "typescript": "^5.4.2"
   }
 }
