@@ -1,14 +1,15 @@
-const { parseFile } = require("music-metadata");
-const fs = require("node:fs/promises");
-const path = require("node:path");
+import { parseFile } from "music-metadata";
+import * as fs from "node:fs/promises";
+import * as path from "node:path";
+
+const songFormats = ['mp3', 'ogg', 'wav', 'mp4', 'webm']
 
 async function main() {
   const publicDir = await fs.readdir("./public");
-
   const items = [];
 
   const tasks = publicDir
-    .filter((item) => item.endsWith(".mp3"))
+    .filter((item) => songFormats.includes(path.extname(item).slice(1)))
     .map(async (item) => {
       const metadata = await parseFile(path.resolve("public", item));
 
@@ -32,6 +33,7 @@ async function main() {
 
   await Promise.all(tasks);
   await fs.writeFile("./music/data.json", JSON.stringify(items));
+  console.log(`${items.length} items written.`)
 }
 
 void main();
